refactor: move Call operation files to compute/gfql/
- Move call_executor.py and call_safelist.py to compute/gfql/ for better organization
- Update all imports to use new location
- All tests pass
