+++ graphistry/compute/validate/validate_schema.py
@@ -331,7 +331,7 @@ def _validate_call_op(
     errors = []
     
     # Import safelist to get schema effects
-    from graphistry.compute.call_safelist import SAFELIST_V1
+    from graphistry.compute.gfql.call_safelist import SAFELIST_V1
     
     # Check if method is in safelist
     if op.function not in SAFELIST_V1:
